backend/requirements.txt
flask
mysql-connector-python
pandas
scikit-learn
google-generativeai
sentence-transformers
torch
requests
PyPDF2
pillow
pytesseract
            